reactive service stub implementation

+++ reactive-kotlin-product-api/src/main/kotlin/org/example/dto/ProductDto.kt
@@ -1,5 +1,5 @@
 package org.example.dto
 
-data class ProductDto(var id: Int,
+data class ProductDto(var id: Long,
                       var name: String,
                       var price: Double)
